Refactor SSH launcher with global backup/restore commands
Major improvements to SSH configuration management:

## 🧹 Clean SSH Launcher
- Removed complex export/import functionality from WezTerm Lua
- Focused SSH launcher on core functionality: connect, add, edit, delete
- Eliminated problematic nested callback issues
- Renamed to clean ssh_launcher.lua (final version)

## 🌍 Global Backup Commands
- Created standalone wezterm_ssh_export.sh with beautiful CLI interface
- Created standalone wezterm_ssh_import.sh with comprehensive validation
- Scripts installed to ~/.local/bin with automatic PATH configuration
- Available globally: just type wezterm_ssh_export.sh anywhere

## 📦 Installer Enhancements
- Automatic script installation to ~/.local/bin
- PATH configuration in .zshrc for global access
- Clear setup instructions for SSH management tools

## 📖 Documentation Updates
- Updated README with global command usage examples
- Added backup/migration section with clear instructions
- Explained benefits of clean separation of concerns
- Added commands to Quick Commands reference table

## 🎯 Benefits
- 🚀 More reliable: no WezTerm callback nesting issues
- 🔧 Better debugging: clear progress feedback and error messages
- 📱 Universal access: works from any terminal anywhere
- 🛡️ Independent: backup tools work regardless of WezTerm changes
- ⚡ Cleaner code: focused responsibilities for each component

Architecture: SSH Launcher handles connections, standalone scripts handle backup/restore
